+++ assets/javascripts/kitten/components/layout/chat-layout/index.js
@@ -72,6 +72,7 @@ ChatLayout.Column = ({
           'k-ChatLayout__column--hasMobileHeader': hasMobileHeader,
         },
       )}
+      tabIndex="-1"
       {...props}
     >
       {isPrimaryColumn && (